refactor(computation): move to main package
- 'computation.New(..)' move to 'hoff.NewComputation(..)'
